refactor: `findProgramInPath` turns out to be unnecessary, so remove it
